Prepending file/line/method to debug logs

+++ blobfuse/utilities.cpp
@@ -67,7 +67,7 @@ void gc_cache::run_gc_cache()
         //check if the closed time is old enough to delete
         if((now - file.closed_time) > file_cache_timeout_in_seconds)
         {
-            syslog(LOG_DEBUG, "File %s being considered for deletion by file cache GC.\n", file.path.c_str());
+            AZS_DEBUGLOGV("File %s being considered for deletion by file cache GC.\n", file.path.c_str());
 
             // path in the temp location
             const char * mntPath;
@@ -85,29 +85,36 @@ void gc_cache::run_gc_cache()
             {
                 //clean up the file from cache
                 int fd = open(mntPath, O_WRONLY);
-                int flockres = flock(fd, LOCK_EX|LOCK_NB);
-                if (flockres != 0)
+                if (fd > 0)
                 {
-                    if (errno == EWOULDBLOCK)
+                    int flockres = flock(fd, LOCK_EX|LOCK_NB);
+                    if (flockres != 0)
                     {
-                        // Someone else holds the lock.  In this case, we will postpone updating the cache until the next time open() is called.
-                        // TODO: examine the possibility that we can never acquire the lock and refresh the cache.
-                        syslog(LOG_DEBUG, "Did not clean up file %s from file cache because there's still an open file handle to it.", mntPath);
+                        if (errno == EWOULDBLOCK)
+                        {
+                            // Someone else holds the lock.  In this case, we will postpone updating the cache until the next time open() is called.
+                            // TODO: examine the possibility that we can never acquire the lock and refresh the cache.
+                            AZS_DEBUGLOGV("Did not clean up file %s from file cache because there's still an open file handle to it.", mntPath);
+                        }
+                        else
+                        {
+                            // Failed to acquire the lock for some other reason.  We close the open fd, and continue.
+                            syslog(LOG_ERR, "Did not clean up file %s from file cache because we failed to acquire the flock for an unknown reason, errno = %d.\n", mntPath, errno);
+                        }
                     }
                     else
                     {
-                        // Failed to acquire the lock for some other reason.  We close the open fd, and continue.
-                        syslog(LOG_ERR, "Did not clean up file %s from file cache because we failed to acquire the flock for an unknown reason, errno = %d.\n", mntPath, errno);
+                        AZS_DEBUGLOGV("GC cleanup of cached file %s.\n", mntPath);
+                        unlink(mntPath);
+                        flock(fd, LOCK_UN);
                     }
+
+                    close(fd);
                 }
                 else
                 {
-                    syslog(LOG_DEBUG, "GC cleanup of cached file %s.\n", mntPath);
-                    unlink(mntPath);
-                    flock(fd, LOCK_UN);
+                    AZS_DEBUGLOGV("Failed to open file %s from file cache in GC, skipping cleanup. errno from open = %d.", mntPath, errno);
                 }
-
-                close(fd);
             }
 
             // lock to remove from front
@@ -136,7 +143,7 @@ int shared_lock_file(int flags, int fd)
             int flockerrno = errno;
             if (flockerrno == EWOULDBLOCK)
             {
-               syslog(LOG_DEBUG, "Failure to acquire flock due to EWOULDBLOCK.  fd = %d.", fd);
+               AZS_DEBUGLOGV("Failure to acquire flock due to EWOULDBLOCK.  fd = %d.", fd);
             }
             else
             {
@@ -175,7 +182,7 @@ int ensure_files_directory_exists_in_cache(const std::string file_path)
         if (slash != pp)
         {
             *slash = '\0';
-            syslog(LOG_DEBUG, "Making cache directory %s.\n", copypath);
+            AZS_DEBUGLOGV("Making cache directory %s.\n", copypath);
             struct stat st;
             if (stat(copypath, &st) != 0)
             {
@@ -209,7 +216,7 @@ std::vector<list_blobs_hierarchical_item> list_all_blobs_hierarchical(std::strin
     int failcount = 0;
     do
     {
-        syslog(LOG_DEBUG, "About to call list_blobs_hierarchial.  Container = %s, delimiter = %s, continuation = %s, prefix = %s\n", container.c_str(), delimiter.c_str(), continuation.c_str(), prefix.c_str());
+        AZS_DEBUGLOGV("About to call list_blobs_hierarchial.  Container = %s, delimiter = %s, continuation = %s, prefix = %s\n", container.c_str(), delimiter.c_str(), continuation.c_str(), prefix.c_str());
 
         errno = 0;
         list_blobs_hierarchical_response response = azure_blob_client_wrapper->list_blobs_hierarchical(container, delimiter, continuation, prefix);
@@ -217,7 +224,7 @@ std::vector<list_blobs_hierarchical_item> list_all_blobs_hierarchical(std::strin
         {
             success = true;
             failcount = 0;
-            syslog(LOG_DEBUG, "Successful call to list_blobs_hierarchical.  results count = %s, next_marker = %s.\n", to_str(response.blobs.size()).c_str(), response.next_marker.c_str());
+            AZS_DEBUGLOGV("Successful call to list_blobs_hierarchical.  results count = %s, next_marker = %s.\n", to_str(response.blobs.size()).c_str(), response.next_marker.c_str());
             continuation = response.next_marker;
             if(response.blobs.size() > 0)
             {
@@ -304,7 +311,7 @@ int is_directory_empty(std::string container, std::string delimiter, std::string
 
 int azs_getattr(const char *path, struct stat *stbuf)
 {
-    syslog(LOG_DEBUG, "azs_getattr called with path = %s\n", path);
+    AZS_DEBUGLOGV("azs_getattr called with path = %s\n", path);
     // If we're at the root, we know it's a directory
     if (strlen(path) == 1)
     {
@@ -329,7 +336,7 @@ int azs_getattr(const char *path, struct stat *stbuf)
     int acc = access(mntPathString.c_str(), F_OK);
     if (acc != -1 )
     {
-        syslog(LOG_DEBUG, "Accessing mntPath = %s for get_attr succeeded; object is in the local cache.\n", mntPathString.c_str());
+        AZS_DEBUGLOGV("Accessing mntPath = %s for get_attr succeeded; object is in the local cache.\n", mntPathString.c_str());
         //(void) fi;
         res = lstat(mntPathString.c_str(), stbuf);
         if (res == -1)
@@ -340,13 +347,13 @@ int azs_getattr(const char *path, struct stat *stbuf)
         }
         else
         {
-            syslog(LOG_DEBUG, "lstat on file %s in local cache succeeded.\n", mntPathString.c_str());
+            AZS_DEBUGLOGV("lstat on file %s in local cache succeeded.\n", mntPathString.c_str());
             return 0;
         }
     }
     else
     {
-        syslog(LOG_DEBUG, "Object %s is not in the local cache during get_attr.\n", mntPathString.c_str());
+        AZS_DEBUGLOGV("Object %s is not in the local cache during get_attr.\n", mntPathString.c_str());
     }
 
     // It's not in the local cache.  Check to see if it's a blob on the service:
@@ -357,7 +364,7 @@ int azs_getattr(const char *path, struct stat *stbuf)
     if ((errno == 0) && blob_property.valid())
     {
 
-        syslog(LOG_DEBUG, "Blob %s, representing a file, found during get_attr.\n", blobNameStr.c_str());
+        AZS_DEBUGLOGV("Blob %s, representing a file, found during get_attr.\n", blobNameStr.c_str());
         stbuf->st_mode = S_IFREG | default_permission; // Regular file (not a directory)
         stbuf->st_uid = fuse_get_context()->uid;
         stbuf->st_gid = fuse_get_context()->gid;
@@ -381,7 +388,7 @@ int azs_getattr(const char *path, struct stat *stbuf)
         }
         if (dirSize != D_NOTEXIST)
         {
-            syslog(LOG_DEBUG, "Directory %s found on the service.\n", blobNameStr.c_str());
+            AZS_DEBUGLOGV("Directory %s found on the service.\n", blobNameStr.c_str());
             stbuf->st_mode = S_IFDIR | default_permission;
             // If st_nlink = 2, means direcotry is empty.
             // Directory size will affect behaviour for mv, rmdir, cp etc.
@@ -393,7 +400,7 @@ int azs_getattr(const char *path, struct stat *stbuf)
         }
         else
         {
-            syslog(LOG_DEBUG, "Entity %s does not exist.  Returning ENOENT (%d) from get_attr.\n", path, ENOENT);
+            AZS_DEBUGLOGV("Entity %s does not exist.  Returning ENOENT (%d) from get_attr.\n", path, ENOENT);
             return -(ENOENT);
         }
     }
@@ -425,7 +432,7 @@ int rm(const char *fpath, const struct stat * /*sb*/, int tflag, struct FTW * /*
 // Delete the entire contents of tmpPath.
 void azs_destroy(void * /*private_data*/)
 {
-    syslog(LOG_DEBUG, "azs_destroy called.\n");
+    AZS_DEBUGLOG("azs_destroy called.\n");
     std::string rootPath(str_options.tmpPath + "/root");
     char *cstr = (char *)malloc(rootPath.size() + 1);
     memcpy(cstr, rootPath.c_str(), rootPath.size());
@@ -480,7 +487,7 @@ int azs_utimens(const char * /*path*/, const struct timespec [2] /*ts[2]*/)
 
 int azs_rename_directory(const char *src, const char *dst)
 {
-    syslog(LOG_DEBUG, "azs_rename_directory called with src = %s, dst = %s.\n", src, dst);
+    AZS_DEBUGLOGV("azs_rename_directory called with src = %s, dst = %s.\n", src, dst);
     std::string srcPathStr(src);
     if (srcPathStr.size() > 1)
     {
@@ -515,7 +522,7 @@ int azs_rename_directory(const char *src, const char *dst)
                 memcpy(&(newDst[dstPathStr.size()]), dir_ent->d_name, nameLen);
                 newDst[dstPathStr.size() + nameLen] = '\0';
 
-                syslog(LOG_DEBUG, "Local object found - about to rename %s to %s.\n", newSrc, newDst);
+                AZS_DEBUGLOGV("Local object found - about to rename %s to %s.\n", newSrc, newDst);
                 if (dir_ent->d_type == DT_DIR)
                 {
                     azs_rename_directory(newSrc, newDst);
@@ -548,7 +555,7 @@ int azs_rename_directory(const char *src, const char *dst)
         return 0 - map_errno(storage_errno);
     }
 
-    syslog(LOG_DEBUG, "Total of %s results found from list_blobs call during rename operation\n.", to_str(listResults.size()).c_str());
+    AZS_DEBUGLOGV("Total of %s results found from list_blobs call during rename operation\n.", to_str(listResults.size()).c_str());
     for (size_t i = 0; i < listResults.size(); i++)
     {
         // We need to parse out just the trailing part of the path name.
@@ -579,7 +586,7 @@ int azs_rename_directory(const char *src, const char *dst)
                 memcpy(&(newDst[dstPathStr.size()]), prev_token_str.c_str(), nameLen);
                 newDst[dstPathStr.size() + nameLen] = '\0';
 
-                syslog(LOG_DEBUG, "Object found on the service - about to rename %s to %s.\n", newSrc, newDst);
+                AZS_DEBUGLOGV("Object found on the service - about to rename %s to %s.\n", newSrc, newDst);
                 if (listResults[i].is_directory)
                 {
                     azs_rename_directory(newSrc, newDst);
@@ -606,7 +613,7 @@ int azs_rename_directory(const char *src, const char *dst)
 // TODO: If/when we upgrade to FUSE 3.0, we will need to worry about the additional possible flags (RENAME_EXCHANGE and RENAME_NOREPLACE)
 int azs_rename(const char *src, const char *dst)
 {
-    syslog(LOG_DEBUG, "azs_rename called with src = %s, dst = %s.\n", src, dst);
+    AZS_DEBUGLOGV("azs_rename called with src = %s, dst = %s.\n", src, dst);
 
     struct stat statbuf;
     errno = 0;
